about 20 hours of codex (#49)
* Implement GameObject systems and related gameplay functions

* Add subsystem stubs for incomplete dependencies

* Restore clang-tidy compatible layout guards

* Make player callback casts host-safe

* Fix remaining cross-platform link and pointer casts

* Improve matching for game object systems

* Checkpoint playable cantina decomp progress

* Port usable gameplay logic from PR #31

* like 20 hours of codex

---------

+++ src/gameapi/edtools/edgra.h
@@ -10,8 +10,8 @@ extern "C" {
     void edGraInitTerrainSwapProtection();
     void edGraEnableTerrainSwap();
     void edGraDisableTerrainSwap();
-    void edgraStartPage(i32 page);
-    void edgraStopPage(i32 page);
+    void edgraStartPage(i8 page);
+    void edgraStopPage(i8 page);
 #ifdef __cplusplus
 }
 #endif
